requirements.txt
requests
python-dotenv
pyspark==3.5.1
delta-spark==3.2.0
pandas
pyarrow>=15.0
pyyaml
folium
altair
geopy